requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
python-pptx
matplotlib